chore: moved types to own file, added toggle button to storybook alias
